# optimize pytorch engine inference with falcon model