--- a/lmdeploy/pytorch/engine/cache_engine.py
+++ b/lmdeploy/pytorch/engine/cache_engine.py
@@ -201,7 +201,8 @@ class CacheEngine:
 
     @staticmethod
     def get_cache_block_size(block_size: int,
-                             model_config: ModelConfig) -> int:
+                             model_config: ModelConfig,
+                             world_size: int = 1) -> int:
         """Get the required cache size of the model.
 
         Args:
@@ -214,6 +215,8 @@ class CacheEngine:
         head_size = model_config.get_head_size()
         num_layers = model_config.num_layers
         num_heads = model_config.num_key_value_heads
+        if not model_config.multi_query_attention:
+            num_heads = num_heads // world_size
 
         key_cache_block = block_size * num_heads * head_size
         value_cache_block = key_cache_block